Initialize the Apollo GraphQL server
This required putting all of the index.ts code into an async function so that the Apollo server could be started asynchonously. The 404 handling logic had to be moved further down. The syntax for exporting various variables had to change since all of the code was now placed in a function instead of the top level.