Adjust character query to work with the GraphQL schema
The type for character data  had to change from the old, manually created type to the new auto-generated type. The TypeORM syntax had to be changed so that the aliases would retain their case sensitivity. This required wrapping all aliases with double quotes. This was something I tried to do but couldn't figure out when initially developing the app as a REST API.