Fix the game data tests
Whenever the function to create the epoch time to expire a key at midnight the next day is called, the mocked data in the tests would increment by one. This resulted in queries for daily records after the expected time range. This changes the day of the month to always be the expected 8th. Additionally, the different tests are isolated by expiring the daily record key